migration

+++ database/migrations/template/2020_10_06_133626_create_contract_items_table.php
@@ -15,7 +15,7 @@ class CreateContractItemsTable extends Migration {
 		Schema::create('template_items', function(Blueprint $table)
 		{
 			$table->increments('id');
-			$table->unsignedBigInteger('template_id')->nullable()->index('contract_template_id');
+			$table->unsignedInteger('template_id')->nullable()->index('contract_template_id');
       $table->text('item');
       $table->timestamps();
 		});
